# improve plugins listing performance