… vendor/golang.org/x/sync/LICENSE
… vendor/golang.org/x/sync/PATENTS
… vendor/golang.org/x/sync/errgroup/errgroup.go
… vendor/modules.txt