follow-up https://github.com/docker/cli/pull/4147#issuecomment-1492232311

**- What I did**

We can slightly improve plugins listing by spawning a goroutine for each iteration.

**- How I did it**

**- How to verify it**

```
$ hyperfine --warmup 2 --runs 5 '/tmp/docker-23.0.2 __complete ""'
Benchmark 1: /tmp/docker-23.0.2 __complete ""
  Time (mean ± σ):      87.7 ms ±   2.4 ms    [User: 66.2 ms, System: 54.2 ms]
  Range (min … max):    83.8 ms …  90.3 ms    5 runs

$ make -f docker.Makefile binary
...

$ hyperfine --warmup 2 --runs 5 './build/docker __complete ""'
Benchmark 1: ./build/docker __complete ""
  Time (mean ± σ):      34.6 ms ±   1.2 ms    [User: 104.4 ms, System: 74.7 ms]
  Range (min … max):    32.8 ms …  36.0 ms    5 runs
```

Also run [this repro](https://github.com/docker/cli/issues/4145).

**- Description for the changelog**
<!--
Write a short (one line) summary that describes the changes in this
pull request for inclusion in the changelog:
-->


**- A picture of a cute animal (not mandatory but encouraged)**